client/client_requirements.txt
pygame==2.4.0
websockets==12.0
PyJWT==2.9.0 #PyJWT==2.8.1
Pillow==10.0.0
numpy==1.24.0
